VFD Constant Calculator
This is a way late commit. These were written a month or more ago. I
hooked up the Vacuum Fluorescent Display to an Arduino Micro and got it
successfully displaying the incremental calculation of E and Pi by
series addition. I'm pleased with it, but got distracted and haven't
polished the project.