Fix "stuck on" condition
I haven't uploaded this to the Arduino yet, but I wanted to get the fix
in before I forget. This *should* fix the issue where pressing the down
button all the way to zero leaves the LEDs stuck on. The problem is the
interruptable delay was causing the shutdown function to return to main
before it actually set the LEDs to zero. With this change the shutdown
function should continue to run no matter if the down button is pushed.
Pushing the up button should cause the shutdown to be interrupted.