Fix long running rainbows
Still haven't uploaded to Arduino to test, but this should fix the fact
that the rainbow sequences stay on far longer than the solid colors. The
color wipes apply pixdelay between each pixel change. For the 60 LED
strip the total time to do one wipe is pixdelay*60. The rainbow sequence
contains the pixdelay in a 0-256 loop and the rainbowCycle contains the
pixdelay in a 0-256*5 loop.  This change divides the delay by those
amounts.