GitHub wonkiness workaround
For some reason GitHub client keeps insisting these files "to be
committed" but shows no changes in the diff and doesn't actually commit
them when I click commit. Added a real change (a comment) in order to
see if that will force GitHub to recognize that they have been
committed.